Add patch method on DistributedCache

+++ api/lib/infrastructure/caches/DistributedCache.js
@@ -15,10 +15,21 @@ class DistributedCache extends Cache {
     this._redisClientSubscriber.on('ready', () => {
       this._redisClientSubscriber.subscribe(this._channel);
     });
-    this._redisClientSubscriber.on('message', () => {
+    this._redisClientSubscriber.on('message', this.clientSubscriberCallback.bind(this));
+  }
+
+  clientSubscriberCallback(message) {
+    // TODO: utiliser un message au format JSON pour le flushall également
+    if (message === 'Flush all') {
       logger.info({ event: 'cache-event' }, 'Flushing the local cache');
       return this._underlyingCache.flushAll();
-    });
+    } else {
+      logger.info({ event: 'cache-event' }, 'Patching the local cache');
+      const parsedMessage = JSON.parse(message);
+      if (parsedMessage.type === 'patch') {
+        this._underlyingCache.patch(parsedMessage.cacheKey, parsedMessage.patch);
+      }
+    }
   }
 
   get(key, generator) {
@@ -27,6 +38,16 @@ class DistributedCache extends Cache {
 
   set(key, object) {
     return this._underlyingCache.set(key, object);
+  }
+
+  patch(key, object) {
+    const message = {
+      patch: object,
+      cacheKey: key,
+      type: 'patch',
+    };
+    const messageAsString = JSON.stringify(message);
+    this._redisClientPublisher.publish(this._channel, messageAsString);
   }
 
   flushAll() {
